Ugh. Updating dog scene animations
there is a lot of crap going on when it comes to the webgl version of the game. I'm getting a lot of errors when building to web, but not when playing in the editor.

So far, in my dog scene, it looks like there are unassigned variables, but I'm double-checking all of my scripts and I can see that they are all assigned in the inspector.

This is an unknown that I really need to figure out.