Updated story 1 intro/outro / fixed a dog error
Story 1: I added animated clouds. It's the dumbest possible way to animate the clouds - they are doubled up so one leaves and one enters the scene and then it loops. Each cloud has it's own animator and the values are hard-coded in. Ugh. It's dumb but it works.

Dog: I had an error with one of the word bubbles not showing up. I wrote a new script to fix that.