Improved Start screen / added reset character button
Added high point to the start screen to jazz it up a bit. Revised character select so that the buttons can be individualized easier, and added the ability to reset the buttons. Also, character select buttons function differently. Clicking the first woman automatically unlocks the man for the next round. clicking the man automatically unlocks the dog. I wrote a new script for it. And duplicated the original scene, so it looks like there's two character select scenes now.